# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgroNutri AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms/{farmId}/recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
    post:
      operationId: createRecommendation
      tags: [recommendations]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
  /farms/{farmId}/recommendations/{recommendationId}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/RecommendationId"
      responses:
        "200":
          description: Recommendation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
    patch:
      operationId: updateRecommendation
      tags: [recommendations]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/RecommendationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
  /farms/{farmId}/recommendations/{recommendationId}/status:
    post:
      operationId: changeRecommendationStatus
      tags: [recommendations]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/RecommendationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationStatusChange"
      responses:
        "200":
          description: Status changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
  /farms/{farmId}/calculations:
    post:
      operationId: runCalculation
      tags: [recommendations]
      summary: Run the deterministic fertigation engine
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculationRequest"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationResult"
...
  /farms/{farmId}/reports:
    get:
      operationId: listReports
      tags: [reports]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  parameters:
    FarmId:
      name: farmId
      in: path
      required: true
      schema:
        type: integer
...
    RecommendationId:
      name: recommendationId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    RecommendationItem:
      type: object
      required: [fertilizerName, weeklyDose, unit]
      properties:
        fertilizerId: { type: ["integer", "null"] }
        fertilizerName: { type: string }
        weeklyDose: { type: number }
        unit: { type: string, description: "kg | L" }
        previousDose: { type: ["number", "null"] }
        reason: { type: ["string", "null"] }
    Recommendation:
      type: object
      required: [id, farmId, status, items, createdAt]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        sectorId: { type: ["integer", "null"] }
        title: { type: ["string", "null"] }
        status: { type: string, description: "draft | pending_review | validated | applying | finished | rejected" }
        source: { type: ["string", "null"], description: "manual | ai" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
        rationale: { type: ["string", "null"] }
        estimatedEcDsM: { type: ["number", "null"] }
        estimatedWeeklyNKg: { type: ["number", "null"] }
        warnings:
          type: array
          items:
            type: string
        createdByName: { type: ["string", "null"] }
        validatedByName: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    RecommendationInput:
      type: object
      required: [items]
      properties:
        sectorId: { type: integer }
        title: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
        rationale: { type: string }
    RecommendationUpdate:
      type: object
      properties:
        title: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
        rationale: { type: string }
    RecommendationStatusChange:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [submit, approve, reject, start_application, finish] }
        comment: { type: string }
    CalculationRequest:
      type: object
      required: [items]
      properties:
        sectorId: { type: integer }
        weeklyLitresPerPlant: { type: number }
        plantCount: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
    CalculationResult:
      type: object
      required: [weeklyWaterLitres, weeklyWaterM3, nutrients, warnings, compatibilityIssues]
      properties:
        weeklyWaterLitres: { type: number }
        weeklyWaterM3: { type: number }
        nutrients:
          type: object
          description: "Weekly kg per nutrient: n, nNitric, nAmmoniacal, nUreic, p2o5, k2o, cao, mgo, so3, b"
          additionalProperties:
            type: number
        estimatedEcDsM: { type: ["number", "null"] }
        waterContribution:
          type: object
          description: "Weekly kg from irrigation water: na, ca, mg, k, b, alkalinity"
          additionalProperties:
            type: number
        sar: { type: ["number", "null"] }
        warnings:
          type: array
          items:
            type: string
        compatibilityIssues:
          type: array
          items:
            type: string
...
    Report:
      type: object
      required: [id, farmId, title, format, status, createdAt]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        title: { type: string }
        format: { type: string, description: "pdf | docx" }
        status: { type: string, description: "generating | ready | error" }
        downloadUrl: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string }
    ReportInput:
      type: object
      required: [format]
      properties:
        title: { type: string }
        format: { type: string, enum: [pdf, docx] }
        recommendationId: { type: integer }
...